+++ restsh/templates/check/http-header/cross-origin-opener-policy.sh
@@ -1,0 +1,23 @@
+# Author: Juergen Mang <juergen.mang@axians.de>
+# Date: 2025-05-12
+
+# Reference: https://cheatsheetseries.owasp.org/cheatsheets/XS_Leaks_Cheat_Sheet.html#cross-origin-opener-policy-coop
+
+read -r VALUE
+
+# Header should be present
+if [ "$VALUE" = "<NOTFOUND>" ]
+then
+    echo "Not found"
+    exit 1
+fi
+
+# OK
+case "$VALUE" in
+    "same-origin-allow-popups") exit 0 ;;
+    "same-origin") exit 0 ;;
+esac
+
+# Invalid value
+echo "Invalid or unsafe value"
+exit 1