+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Feat: Add `f5.asm.policy.signature.list`
 - Feat: Pressing `Ctrl`+`h` shows the help of current command
+- Feat: More options for `restsh.check.http-header`
 
 ## v4.1.0 - 2025-04-25